Added fix for more correct offset.

+++ kafka/include/userver/kafka/offset_range.hpp
@@ -10,11 +10,11 @@ namespace kafka {
 struct OffsetRange final {
     /// @brief The low watermark offset. It indicates the earliest available offset in Kafka.
     /// @note low offset is guaranteed to be committed
-    std::uint32_t low{};
+    std::int64_t low{};
 
     /// @brief The high watermark offset. It indicates the next offset that will be written in Kafka.
     /// @note high offset is not required to be committed yet
-    std::uint32_t high{};
+    std::int64_t high{};
 };
 
 }  // namespace kafka
