+++ kafka/src/kafka/impl/consumer_impl.cpp
@@ -152,7 +152,8 @@ void PrintTopicPartitionsList(
     }
 
     const utils::span<const rd_kafka_topic_partition_t> topic_partitions{
-        list->elems, list->elems + static_cast<std::size_t>(list->cnt)};
+        list->elems, list->elems + static_cast<std::size_t>(list->cnt)
+    };
     for (const auto& topic_partition : topic_partitions) {
         if (skip_invalid_offsets && topic_partition.offset == RD_KAFKA_OFFSET_INVALID) {
             /// @note `librdkafka` does not sets offsets for partitions that were
@@ -355,7 +356,7 @@ OffsetRange ConsumerImpl::GetOffsetRange(
         throw OffsetRangeException{fmt::format("Failed to get offsets: invalid offset."), topic, partition};
     }
 
-    return {static_cast<std::uint32_t>(low_offset), static_cast<std::uint32_t>(high_offset)};
+    return {low_offset, high_offset};
 }
 
 std::vector<std::uint32_t>
@@ -380,7 +381,8 @@ ConsumerImpl::GetPartitionIds(const std::string& topic, std::optional<std::chron
     }()};
 
     const utils::span<const rd_kafka_metadata_topic> topics{
-        metadata->topics, static_cast<std::size_t>(metadata->topic_cnt)};
+        metadata->topics, static_cast<std::size_t>(metadata->topic_cnt)
+    };
     const auto* topic_it =
         std::find_if(topics.begin(), topics.end(), [&topic](const rd_kafka_metadata_topic& topic_raw) {
             return topic == topic_raw.topic;
@@ -390,7 +392,8 @@ ConsumerImpl::GetPartitionIds(const std::string& topic, std::optional<std::chron
     }
 
     const utils::span<const rd_kafka_metadata_partition> partitions{
-        topic_it->partitions, static_cast<std::size_t>(topic_it->partition_cnt)};
+        topic_it->partitions, static_cast<std::size_t>(topic_it->partition_cnt)
+    };
     std::vector<std::uint32_t> partition_ids;
     partition_ids.reserve(partitions.size());
